refactor: admin page 관련 ui 상태 app에서 분리
