+++ src/basic/App.tsx
@@ -1,6 +1,5 @@
 import { useState, useCallback, useEffect } from 'react';
 
-import { Product } from '../types';
 import { AdminPage } from './components/AdminPage';
 import { CartPage } from './components/CartPage';
 import { Header } from './components/Header';
@@ -9,11 +8,6 @@ import { useCart } from './hooks/useCart';
 import { useCoupons } from './hooks/useCoupons';
 import { useNotifications } from './hooks/useNotifications';
 import { useProducts } from './hooks/useProducts';
-
-interface ProductWithUI extends Product {
-  description?: string;
-  isRecommended?: boolean;
-}
 
 const App = () => {
   const { notifications, setNotifications, addNotification } = useNotifications();
@@ -41,28 +35,9 @@ const App = () => {
   });
 
   const [isAdmin, setIsAdmin] = useState(false);
-  const [showCouponForm, setShowCouponForm] = useState(false);
-  const [activeTab, setActiveTab] = useState<'products' | 'coupons'>('products');
-  const [showProductForm, setShowProductForm] = useState(false);
+
   const [searchTerm, setSearchTerm] = useState('');
   const [debouncedSearchTerm, setDebouncedSearchTerm] = useState('');
-
-  // Admin
-  const [editingProduct, setEditingProduct] = useState<string | null>(null);
-  const [productForm, setProductForm] = useState({
-    name: '',
-    price: 0,
-    stock: 0,
-    description: '',
-    discounts: [] as Array<{ quantity: number; rate: number }>,
-  });
-
-  const [couponForm, setCouponForm] = useState({
-    name: '',
-    code: '',
-    discountType: 'amount' as 'amount' | 'percentage',
-    discountValue: 0,
-  });
 
   const formatPrice = (price: number, productId?: string): string => {
     if (productId) {
@@ -91,45 +66,6 @@ const App = () => {
     clearCart();
   }, [addNotification]);
 
-  const handleProductSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    if (editingProduct && editingProduct !== 'new') {
-      updateProduct(editingProduct, productForm);
-      setEditingProduct(null);
-    } else {
-      addProduct({
-        ...productForm,
-        discounts: productForm.discounts,
-      });
-    }
-    setProductForm({ name: '', price: 0, stock: 0, description: '', discounts: [] });
-    setEditingProduct(null);
-    setShowProductForm(false);
-  };
-
-  const handleCouponSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    addCoupon(couponForm);
-    setCouponForm({
-      name: '',
-      code: '',
-      discountType: 'amount',
-      discountValue: 0,
-    });
-    setShowCouponForm(false);
-  };
-
-  const startEditProduct = (product: ProductWithUI) => {
-    setEditingProduct(product.id);
-    setProductForm({
-      name: product.name,
-      price: product.price,
-      stock: product.stock,
-      description: product.description || '',
-      discounts: product.discounts || [],
-    });
-    setShowProductForm(true);
-  };
 
   const totals = calculateCartTotal();
 
@@ -155,32 +91,16 @@ const App = () => {
       <main className='max-w-7xl mx-auto px-4 py-8'>
         {isAdmin ? (
           <AdminPage
-            // --- 탭 관련 상태 ---
-            activeTab={activeTab}
             // --- 데이터 엔티티 ---
             products={products}
             coupons={coupons}
-            // --- UI 및 form 관련 상태 ---
-            editingProduct={editingProduct}
-            showProductForm={showProductForm}
-            productForm={productForm}
-            showCouponForm={showCouponForm}
-            couponForm={couponForm}
-            // --- 탭 관련 핸들러 ---
-            setActiveTab={setActiveTab}
             // --- 상품 관련 핸들러 ---
+            addProduct={addProduct}
+            updateProduct={updateProduct}
             deleteProduct={deleteProduct}
-            startEditProduct={startEditProduct}
-            handleProductSubmit={handleProductSubmit}
             // --- 쿠폰 관련 핸들러 ---
+            addCoupon={addCoupon}
             deleteCoupon={deleteCoupon}
-            handleCouponSubmit={handleCouponSubmit}
-            // --- form 관련 상태 변경 함수 ---
-            setEditingProduct={setEditingProduct}
-            setShowProductForm={setShowProductForm}
-            setProductForm={setProductForm}
-            setShowCouponForm={setShowCouponForm}
-            setCouponForm={setCouponForm}
             // --- 유틸 함수 ---
             formatPrice={formatPrice}
             addNotification={addNotification}
